add ability for product catalog service to fail selectively based on the pod name
This feature allows you to demo a scenario where a single pod is failing - e.g. because it a new dpeloyment with a new service.version that introduces a bug

This change is supported to be upstreamed, together with changes in helm charts that create copies of the product catalog service with different pod names (e.g. foo-productcatalogservice2-bar, foo and bar represent the other name parts in k8s that we're not interested in)